requirements.txt
kafka-python==2.2.10
pyspark==3.4.1
